New datafiles for the share/msgdef collection.
The data files contain the unicode encoded character for the locale
msgdef data. This can be converted to either UTF-8 (which will be
the main file) and symlinked where possible or to other charactermaps.

This way we don't end up with high-ascii or otherwise encoded data
files in the repository and we will be able to edit the files without
having to worry about the right characters because they later get
converted into the right encoding.