Pf(4) needs a mechanism to be notified that the destination address changed after pf(4) has seen the packet on input path and might see it again in the output pathi(state match optimization).
Since this needs to touch many subsystems implement a wrapper
that will call all callbacks for other subsystems that might see
this information useful.
Basically they just place the callback in this wrapper rather than
go through all the sources that this function will be.

Idea and basic template from mlaier@.

NOTE: With this commit pf 4.5 can be considered ported to FreeBSD with
all its features.