Remove a hack of questionable continuing relevance, though checks will be needed on all architectures before this can go into HEAD.
The ELF image activator used the segment containing the image entry point
as the official text segment, and the last other loadable segment as the
official data segment, instead of checking whether the segments were
executable. This was apparently required for Alpha support. On PPC64,
the executable entry point points to a non-executable function
descriptor in the .opd data section, which breaks this logic. As we no
longer support Alpha, it may now be safe to turn off this hack.