Start to merge functionality from NetBSD.  Merge the ability to specify the full path instead of the relative path.  This brings in all of NetBSD's spec.c, with other bits and pieces to support it.
This is a work in progress.  At the very least we need to cut over to
mtree_err in a few more places.  We need to bring in the merge code as
well, so we can merge the BSD.foo.mtree files with the metadata files
to allow the nopriv build to use makefs to build images.  mtree gets
cranky if you don't list all the dirs before files in the dirs...