Working towards making the cryptocteon code compile and work on FreeBSD.  Some convoluted use of scatterlists has been replaced with direct use of pointers into a linear buffer since it appeared that only a single scatterlist entry was being used; I intend to make that use iovecs instead in the near future.
The crypto device front-end is mostly commented out and is not compiled.

Actual enabling and disabling of use of the crypto coprocessor within the
kernel is unimplemented in this commit, but is trivial for kernel-only use.
We need a mechanism that allows use of the coprocessor from userland at some
point, but that's a stretch goal for me.