This is a small program that runs a series of performance tests designed to reveal the physical block size of a disk.  It performs a series of staggered reads and / or writes of different lengths at different alignments.  The idea is that a disk that has a large physical blocksize but simulates a smaller one will perform poorly, due to read-modify-write, except when both the I/O size and the offset are multiples of the physical block size.
The program supports read-only, write-only and read-write tests.  The
default is write-only.  A read-write test is probably not very useful,
because it primes the cache in cases where the disk would actually
perform an extra read.  A read-only test is useless except as a quick
way to verify that the program works.

Note that the program currently assumes a logical block size of 512
bytes and that the parameters (size range, interval and iterations)
are hardcoded.