(re-)add another rendezvous counter
The problem was that if a master cpu calls rendezvous in rapid succession
then the rendezvous object could be re-used while some slave cpus were
still spinning on the dual-function teardown/exit counter if a teardown
function was actually specified.
To fix this problem use separate counters, one for teardown entry and the
other for signaling full rendezvous completion.