Remove the need for routes to have the MTU explicitly set or inherited from their interface.  Unless explicitly set the route MTU is zero and the current interface MTU is being used.
This allows interface MTU to be increased even after it got an IP address
assigned.

If a route has an MTU set that is lower than the interface MTU it will
be used instead.  If a route has an MTU set that is higher than the
interface MTU it will be adjusted and the interface MTU will be used
as before.