Replacement toedev/toecore interface.
Interfaces in the toedev struct are all the hooks a TOE driver exposes
to the kernel (toe_usrreqs etc. will all be retired).  A few helper
functions for use by a TOE driver will also be added to this module.
All TOE drivers are expected to have a dependency on the toecore module.