Make sure that global locks have their own CPU cache line in the .bss section of the kernel and do not end up sharing one by chance or the place they are defined in.
Three new macros are added:
 MTX_GLOBAL(name)		defining the mutex
 MTX_GLOBAL_STATIC(name)	defining the mutex as static
 MTX_GLB(name)			accessing the mutex in mtx_lock()

This makes global locks cache deterministic.  Changes to the
.bss layout and ordering due to differences in completely
unrelated parts of the kernel can no longer cause global
locks to share the same CPU cache line.

This is a proof of concept with only one global lock converted.

I'm open to suggestions and improvements on the macros and
their naming.