Merge ixgbe_tx_ctx_setup() and ixgbe_tso_setup() together into ixgbe_offload_setup() as they have a large overlap in packet inspection and variables.
Add UDP fragmentation offload functionality as well.

Also the driver can assume that the necessary headers for the
activated offload functions are contiguously present in the
first mbuf.  This assumption is not formalized yet but significantly
simplify the driver code.  The upper layers of the stack adhere
to this assumption as well.  We leave enough leading space in the
mbufs to prepend all registered (max_linkhdr) encapsulations.
There may be misbehaving packet transformations that have to be
fixed.  Assertion functions have to be provided as well.

This change is not tested yet because I lack the hardware.
This change translates well to other driver with a similar structure.