Remove IN_ZERONET macro and test from icmp_reflect() that was added with r120958 to drop such packets.  It was only applied to ICMP while the normal ip_input() path continues to allow it leading to inconsistent behavior between ICMP and TCP/UDP.
Due to the global IPv4 address shortage it may be helpful to
have 0.0.0.0/8, with the exception of 0.0.0.0/32, available as
another private IP range.

A close reading of RFC5735 and RFC1122 is not entirely conclusive.
While usage in a local ("this") scope may be allowed, forwarding
may be prohibited.

A survey on how other OS's handle 0.0.0.0/8 should be conducted.

Reported by:	Sean Chittenden <sean-at-chittenden.org>