Extend the checksum offload fields in the mbuf packet header with layer 2-4 header length indicators for the current packet. It is only valid on the down/outbound path with the corresponding csum flags specified.
The header length indicators are 8-bit wide and can specify a
length of at most 256 bytes per header which is deemed sufficient.

The header length indicators are unionized with the csum_data
field.

This simplifies setup of the offload DMA desriptors in the
drivers considerably as they don't have to parse the packet
anymore.

Adjustments to Ethernet/IP/IPv6/UDP/TCP/SCTP to correctly set
and update the header length indicators to follow.