Add sysctl support for pfil hook reporting and ordering.
pfil_head_register() takes an additional pointer to the parent sysctl
node pointer of the protocol family that is being registered here.

Two new branches are then attached named "pfil_in" and "pfil_out".
In each branch a list of the attached pfil hooks in the order of
their processing is provided in the "hooks" string sysctl.

In addition each attached pfil hook has its own node with a numeric
representation of is ordering rank (0-255).  Writing a new ordering
rank changes the processing order.  If two hooks have the same rank
the last installed or changed comes first.

Example output:
 net.inet.pfil_in.hooks="pf,ipfw,ipfilter"
 net.inet.pfil_in.pf=200
 net.inet.pfil_in.ipfw=201
 net.inet.pfil_in.ipfilter=202
 net.inet.pfil_out.hooks="ipfilter,ipfw,pf"
 net.inet.pfil_out.ipfilter=53
 net.inet.pfil_out.ipfw=54
 net.inet.pfil_out.pf=55

This may be further tweaked and refined.