Rework copy routines to ensure we always use memory allocated via EFI.
The previous code assumed it could copy wherever it liked. This is not the
case. The approach taken by this code is pretty ham-fisted in that it simply
allocates a large (32MB) buffer area and stages into that, then copies the
whole area into place when it's time to execute. A more elegant solution could
be used but this works for now.

Sponsored by:	FreeBSD Foundation