Adjust our load device when we boot from CD under UEFI.
The process for booting from a CD under UEFI involves adding a FAT filesystem
containing your loader code as an El Torito boot image. When UEFI detects
this, it provides a block IO instance that points at the FAT filesystem as a
child of the device that represents the CD itself. The problem being that the
CD device is flagged as a "raw device" while the boot image is flagged as a
"logical partition". The existing EFI partition code only looks for logical
partitions and so the CD filesystem was rendered invisible.

To fix this, check the type of each block IO device. If it's found to be a
CD, and thus an El Torito boot image, look up its parent device and add that
instead so that the loader will then load the kernel from the CD filesystem.
This is done by using the handle for the boot filesystem as an alias.

Something similar to this will be required for booting from other media as well
as the loader will live in the EFI system partition, not on the partition
containing the kernel.