Add peer and key handling for listen and connected sockets as well as the setsockopt() key interface.
When TCP-AO is activated on a TCP socket a tcp_ao control block
is allocated and attached to the tcpcb.  An arbitrary number of
peer addresses and keys can be added.  On accepting of a new
incoming connection only the peers keys are copied to the new
socket.  On outbound connecting sockets only the keys for the
peer address will be kept.

Some parts are not fully implemented yet and other may change
again.

Sponsored by:	Juniper Networks