Generalize the bitarray concept implemented in cpuset and implement a KPI to make it reusable: bitset. bitsets represent arrays of arch's registers sized entries and offer accessors function to work with them.
All the consumers may want to specify a new set by using BITSET_DEFINE()
KPI and specify some of the needed operations.

For how it is intrisically implemented, kern_cpuset.c and sparc64 KTR
support are aware of internal details of such implementation, but they
can be changed in the future to deal with generic containers as well
if this is assumed to be beneficial and easier.

Sponsored by:	EMC / Isilon storage division
Obtained from:	jeff