MFV r254749:
Don't hold dd_lock for long by breaking it when not doing dsl_dir
accounting.  It is not necessary to hold the lock while manipulating
the parent's accounting, because there is no interface for userland
to see a consistent picture of both parent and child at the same
time anyway.

Illumos ZFS issues:
  4046 dsl_dataset_t ds_dir->dd_lock is highly contended


git-svn-id: svn+ssh://svn.freebsd.org/base/head@254757 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f