Improve the logic used to determine the default value of the HOSTNAME configuration variable.
Introduce an NCPU configuration variable and implement similar logic
as for the HOSTNAME configuration variable.

Change the configuration variable substitution syntax from %%FOO%% to
${FOO}.  The old syntax is still supported for backward compatibility,
but is not used internally and is deemphasized in the documentation.

Introduce support for using environment variables in configuration
files.  This requires a fair amount of validation and untainting.

Clean up the man page, especially the list of configuration variables,
and provide more detailed explanations of some of the darker corners
of the code as well as advice about tuning JOBS and NCPU.