Add scripts to set up a single Kyua testing node.
This new 'autotest' directory is intended to contain scripts and
configuration files used to set up the Kyua testing cluster.  This
specific commit brings in only the code used to bring up a single
node in the cluster.

Be aware that most of the code in here is only temporary glue until Kyua
itself becomes a better test harness with support for multiple hosts and
consolidated reports.  (All this stuff is planned but it's a non-trivial
amount of work so, in the meantime, we get something ad hoc but useful.)

It is also possible that parts of this code might end up in src (like the
generation of VM disk images).  However, this whole thing is not intended
to be part of src and is why it is being checked in the user/ hierarchy.

And, finally: what's in the autotest name?  Nothing special.