Rename the rc.d script to autotest_node.
I don't think dashes are easily supported as the 'name' variable of an
rc.d script, and the inconsistency between the file name and the rc.conf
variables was confusing.  (Actually, the setup script was broken because
it attempted to start the service using an underscore.)

To prevent confusion, just rename the script to autotest_node.