As result of a number of inquiries, add the release.sh wrapper script to my user directory for public consumption.
Thermite is a pyrotechnic composition of shell script and zfs.
When executed, it generates a significant amount of heat.

Here's how it works:

 - Create the configuration files for each build to be run.
   The file name format is: {rev}-{arch}-{kernel}-{type}, where
   rev is the revision to build (i.e., 10, 9), arch is the
   TARGET_ARCH to be built, kernel is the KERNCONF, and type
   is the type of build ('snap' or 'release').  An example is
   included as 11-amd64-GENERIC-snap.conf.

 - Edit master.conf for the build environment, in particular,
   make sure zfs_root and zfs_mount are updated if the builder
   is a zfs-enabled machine.  If the builder is zfs-enabled,
   a number of redundant steps are skipped, and 'zfs clone' is
   used to shorten overall times.

 - Run:
	# ./thermite.sh -c ./master.conf

 - Cross your fingers.

The mk-vmimage.sh script uses qemu-img to create the VM images
that are published on ftp.FreeBSD.org.  It is also included, as
parts of thermite.sh call it after successful builds.

The master.conf file is effectively what is used for the
weekly snapshot builds.

Sponsored by:	The FreeBSD Foundation