Turn setup into an shtk-based script.
This means that "pkg install shtk" must be run first on the host by hand,
but that's easy enough to do.  Added a check to the Makefile to ensure
that shtk is present.

The reason for doing this is that I want to add extra logic to this script,
and that logic will be much easier to implement with shtk.