Make setup.sh allow a controlled deployment of specific autotest versions.
With this change, setup.sh will fetch a separate copy of autotest (and its
dependencies!) based on a revision specification in a per-host
configuration file.  This separate copy is the one used on the worker node
to run the tests.

Additionally, setup.sh will register itself as a cron job to maintain the
autotest code and all configuration files up-to-date.

The goal of this change is to permit having one of the kyua[123] workers
as a canary of configuration changes while the others remain stable.  This
is much necessary as I've been known to routinely break the nodes when I
manually push to them untested changes.