Enable ARM_INTRNG on IMX6 platforms, and make the imx_gpio driver an interrupt controller.
The latter is required for INTRNG, because of the hardware erratum
workaround installed by the linux folks into the imx6 FDT data, which remaps
an ethernet interrupt to the gpio device.  In the non-INTRNG world we
intercept the call to map the interrupt and map it back to the ethernet
hardware (because we don't need linux's workaround), but in the INTRNG world
we lose the hookpoint where that remapping was happening, but we gain the
ability to work the way linux does by having the gpio driver dispatch the
interrupt.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@289537 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f