ath(4) - don't try to free buffers / return an error if we've committed to transmit the buffer.
ath_tx_start() may manipulate/reallocate the mbuf as part of the DMA
code, so we can't expect the mbuf can be returned back to the caller.
Now, the net80211 ifnet work changed the semantics slightly so
if an error is returned here, the mbuf/reference is freed by the
caller (here, it's net80211.)

So, once we reach ath_tx_start(), we never return failure.  If we fail
then we still return OK and we free the mbuf/noderef ourselves, and
we increment OERRORS.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@290339 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f