Reintegrate colldef(1) and mklocale(1)
While those tools are not needed anymore they are necessary to build FreeBSD 9
and 10. it does not hurt to keep those tools around until both 9 and 10 branch
become EOLed.

Modify colldef(1) to build after the change in the collation header, and ensure
it does produce the same collation definition it used to generate for 9 and 10

Reported by:	Oliver Pinter


git-svn-id: svn+ssh://svn.freebsd.org/base/head@291115 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f