hyperv/hn: Make # of rings configurable
And since the host may not being able to allocate the # of rings
requested by us, save the # of rings allocated by the host in the
ring_inuse counters; use ring_inuse counters for run time operation.

This paves the way for the upcoming vRSS support.

MFC after:	1 week
Sponsored by:	Microsoft OSTC


git-svn-id: svn+ssh://svn.freebsd.org/base/head@296305 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f