Provide working examples which use PACKAGE/${PACKAGE}FILES
file1 was previously created on the fly and deleted after the test was
over. Switch things up a bit by installing it via the build, so helpful
comments and reasoning can be added for the ${PACKAGE}FILES idiom I've
sprinkled around the tree on the branch

Also, in plain/cp_test.sh, use file3 in the second test instead of file2.
Using file2 worked previously because the files' permissions were reset
when the files were truncated, but since the environment for both the
`force` and `simple` testcases are the same, it would fail when `force`
followed `simple`, as file2 existed when `force` was run, but had
incorrect permissions, thus resulting in a permission died error from
cp(1).