makefs(1): use all the random(3) range.
The generation number is uint32_t so we can fit the complete range
of random(3). We could have used arc4random() but the result would
be unpredictable and it would prohibit reproducible builds.

While here add a comment where seeding is done: this affects
reproducible builds and might have to be re-visited to use a
release dependent value.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@300069 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f