Fix the signature of the psignal() function.
POSIX 2008 added the psignal() function which has already been part of
the BSDs for a long time. The only difference is, the POSIX version uses
an 'int' for the signal number, unlike our version which uses an
'unsigned int'. Fix up the function to use an 'int'. This should not
affect the ABI.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@300997 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f