Merge {amd64,i386}/instr_size.c into x86_instr_size.c.
Also reduce the diff between us and upstream: the input data model will
always be DATAMODEL_NATIVE because of a bug (p_model is never set but is
always initialized to 0), so we don't need to override the caller anyway.
This change is also necessary to support the pid provider for 32-bit
processes on amd64.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@303050 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f