Add dumpon(8) support for netdump client configuration.
In general we need four configuration parameters: client/server/gateway
addresses, and the name of the interface to be used. dumpon is modified
to accept (and in some cases, infer) these parameters and uses ioctls
on /dev/netdump to pass them to the client code.

This commit is lacking corresponding man page updates.