Add vm_page_alloc_pages_after().
This is a new page allocation which is intended to complement
vm_page_grab_pages(). It permits the allocation of multiple pages,
contiguous in the pindex space, with a single call. When
VM_ALLOC_{NOWAIT,WAITFAIL} are specified, the returned run may be
shorter than the one requested. In support of this function,
vm_reserv_extend() and vm_reserv_alloc_page() may now optionally
return a run of contiguous pages from the same reservation, and the
new vm_phys_alloc_npages() function is used to allocate pages from
the physical memory allocator.